requirements.txt
flask==2.3.3
Werkzeug==2.3.7
flask-socketio==5.3.6
python-socketio==5.9.0
numpy==1.25.2
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
ultralytics==8.0.196
eventlet==0.33.3
python-engineio==4.5.1
gevent==23.9.1
gevent-websocket==0.10.1
pillow==10.0.1
protobuf==4.24.3
tqdm==4.66.1
requests==2.31.0 